Add bls feature for optionally disabling it
