+++ monad-compress/Cargo.toml
@@ -14,8 +14,8 @@ lz4 = { workspace = true }
 
 [dev-dependencies]
 monad-consensus = { path = "../monad-consensus" }
-monad-consensus-types = { path = "../monad-consensus-types" }
-monad-crypto = { path = "../monad-crypto", features = ["secp"] }
+monad-consensus-types = { path = "../monad-consensus-types", features = ["bls"] }
+monad-crypto = { path = "../monad-crypto", features = ["bls", "secp"] }
 monad-state = { path = "../monad-state" }
 monad-testutil = { path = "../monad-testutil" }
 monad-types = { path = "../monad-types" }
