+++ monad-consensus-types/Cargo.toml
@@ -26,7 +26,7 @@ zerocopy = { workspace = true }
 
 
 [dev-dependencies]
-monad-crypto = { path = "../monad-crypto", features = ["secp"] }
+monad-crypto = { path = "../monad-crypto", features = ["bls", "secp"] }
 monad-testutil = { path = "../monad-testutil" }
 
 criterion = { workspace = true }
@@ -34,6 +34,10 @@ rand = { workspace = true }
 rand_chacha = { workspace = true }
 test-case = { workspace = true }
 
+[features]
+bls = ["monad-crypto/bls"]
+
 [[bench]]
 name = "bls"
 harness = false
+required-features = ["bls"]
