+++ monad-consensus-types/src/bls.rs
@@ -9,7 +9,7 @@ use monad_crypto::{
     certificate_signature::PubKey,
     hasher::{Hash, Hashable, Hasher, HasherType},
 };
-use monad_proto::proto::signing::ProtoBlsSignatureCollection;
+use monad_proto::{error::ProtoError, proto::signing::ProtoBlsSignatureCollection};
 use monad_types::NodeId;
 use prost::Message;
 
@@ -274,6 +274,31 @@ impl<PT: PubKey> SignatureCollection for BlsSignatureCollection<PT> {
             .map_err(|e| SignatureCollectionError::DeserializeError(format!("{}", e)))?;
         bls.try_into()
             .map_err(|e| SignatureCollectionError::DeserializeError(format!("{}", e)))
+    }
+}
+
+impl<PT: PubKey> From<&BlsSignatureCollection<PT>> for ProtoBlsSignatureCollection {
+    fn from(value: &BlsSignatureCollection<PT>) -> Self {
+        Self {
+            signers: serde_cbor::to_vec(&value.signers)
+                .expect("serialization success")
+                .into(),
+            sig: value.sig.serialize().into(),
+        }
+    }
+}
+
+impl<PT: PubKey> TryFrom<ProtoBlsSignatureCollection> for BlsSignatureCollection<PT> {
+    type Error = ProtoError;
+
+    fn try_from(value: ProtoBlsSignatureCollection) -> Result<Self, Self::Error> {
+        Ok(Self {
+            signers: serde_cbor::from_slice(&value.signers)
+                .map_err(|e| ProtoError::DeserializeError(format!("{}", e)))?,
+            sig: BlsAggregateSignature::deserialize(&value.sig)
+                .map_err(|e| ProtoError::CryptoError(format!("{}", e)))?,
+            _phantom: PhantomData,
+        })
     }
 }
