+++ monad-consensus-types/src/convert/signing.rs
@@ -1,13 +1,7 @@
-use std::marker::PhantomData;
-
-use monad_crypto::{
-    bls12_381::BlsAggregateSignature,
-    certificate_signature::{CertificateSignature, CertificateSignatureRecoverable, PubKey},
-};
+use monad_crypto::certificate_signature::{CertificateSignature, CertificateSignatureRecoverable};
 use monad_proto::{error::ProtoError, proto::signing::*};
 
 use crate::{
-    bls::BlsSignatureCollection,
     multi_sig::MultiSig,
     signature_collection::{SignatureCollection, SignatureCollectionError},
 };
@@ -34,31 +28,6 @@ impl<S: CertificateSignatureRecoverable> TryFrom<ProtoMultiSig> for MultiSig<S> 
                 .into_iter()
                 .map(|v| proto_to_certificate_signature(v))
                 .collect::<Result<Vec<_>, _>>()?,
-        })
-    }
-}
-
-impl<PT: PubKey> From<&BlsSignatureCollection<PT>> for ProtoBlsSignatureCollection {
-    fn from(value: &BlsSignatureCollection<PT>) -> Self {
-        Self {
-            signers: serde_cbor::to_vec(&value.signers)
-                .expect("serialization success")
-                .into(),
-            sig: value.sig.serialize().into(),
-        }
-    }
-}
-
-impl<PT: PubKey> TryFrom<ProtoBlsSignatureCollection> for BlsSignatureCollection<PT> {
-    type Error = ProtoError;
-
-    fn try_from(value: ProtoBlsSignatureCollection) -> Result<Self, Self::Error> {
-        Ok(Self {
-            signers: serde_cbor::from_slice(&value.signers)
-                .map_err(|e| ProtoError::DeserializeError(format!("{}", e)))?,
-            sig: BlsAggregateSignature::deserialize(&value.sig)
-                .map_err(|e| ProtoError::CryptoError(format!("{}", e)))?,
-            _phantom: PhantomData,
         })
     }
 }
