+++ monad-crypto/Cargo.toml
@@ -12,7 +12,7 @@ bench = false
 monad-proto = { path = "../monad-proto" }
 
 blake3 = { workspace = true }
-blst = { workspace = true }
+blst = { workspace = true, optional = true }
 secp256k1 = { workspace = true, features = ["global-context", "recovery"], optional = true }
 sha2 = { workspace = true }
 zerocopy = { workspace = true }
@@ -29,6 +29,7 @@ hex = { workspace = true }
 tiny-keccak = { workspace = true, features = ["keccak"] }
 
 [features]
+bls = ["dep:blst"]
 rustls = ["dep:ring", "dep:rcgen", "dep:rustls", "dep:x509-parser"]
 secp = ["dep:secp256k1"]
