+++ monad-node/Cargo.toml
@@ -12,8 +12,8 @@ bench = false
 
 [dependencies]
 monad-consensus-state = { path = "../monad-consensus-state" }
-monad-consensus-types = { path = "../monad-consensus-types" }
-monad-crypto = { path = "../monad-crypto", features = ["secp"] }
+monad-consensus-types = { path = "../monad-consensus-types", features = ["bls"] }
+monad-crypto = { path = "../monad-crypto", features = ["bls", "secp"] }
 monad-eth-types = { path = "../monad-eth-types", features = ["serde"] }
 monad-executor = { path = "../monad-executor" }
 monad-executor-glue = { path = "../monad-executor-glue" }
