+++ monad-virtual-bench/Cargo.toml
@@ -12,7 +12,7 @@ bench = false
 
 [dev-dependencies]
 monad-consensus-state = { path = "../monad-consensus-state" }
-monad-consensus-types = { path = "../monad-consensus-types" }
+monad-consensus-types = { path = "../monad-consensus-types", features = ["bls"] }
 monad-crypto = { path = "../monad-crypto" }
 monad-executor = { path = "../monad-executor" }
 monad-executor-glue = { path = "../monad-executor-glue" }
